WIP using interface struct
